+++ crates/cheatnet/src/cheatcodes.rs
@@ -11,7 +11,7 @@ use thiserror::Error;
 pub mod declare;
 pub mod deploy;
 pub mod get_class_hash;
-pub mod l1_handler_call;
+pub mod l1_handler_invoke;
 pub mod prank;
 pub mod roll;
 pub mod warp;